% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{autoplot.hclust}
\alias{autoplot.linkage}
\alias{autoplot.dendrogram}
\alias{autoplot.agnes}
\alias{autoplot.diana}
\alias{autoplot.twins}
\alias{autoplot.pvclust}
\title{Objects exported from other packages}
\usage{
\method{autoplot}{hclust}(object, layout = "dendrogram", ladderize = FALSE, hang = 0.1, ...)

\method{autoplot}{linkage}(object, layout = "dendrogram", ladderize = FALSE, hang = 0.1, ...)

\method{autoplot}{dendrogram}(object, layout = "dendrogram", ladderize = FALSE, hang = 0.1, ...)

\method{autoplot}{agnes}(object, layout = "dendrogram", ladderize = FALSE, hang = 0.1, ...)

\method{autoplot}{diana}(object, layout = "dendrogram", ladderize = FALSE, hang = 0.1, ...)

\method{autoplot}{twins}(object, layout = "dendrogram", ladderize = FALSE, hang = 0.1, ...)

\method{autoplot}{pvclust}(
  object,
  layout = "dendrogram",
  ladderize = FALSE,
  label_edge = FALSE,
  pvrect = FALSE,
  alpha = 0.95,
  hang = 0.1,
  ...
)
}
\arguments{
\item{object}{input object}

\item{layout}{layout for plotting the tree}

\item{ladderize}{whether ladderize the tree (default FALSE)}

\item{hang}{numeric The fraction of the tree plot height by which labels
should hang below the rest of the plot. A negative value will cause the
labels to hang down from 0.}

\item{...}{additional paramters that passed to ggtree}
}
\value{
ggtree object
}
\examples{
d <- dist(USArrests)
hc <- hclust(d, "ave")
autoplot(hc) + geom_tiplab()
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}
}}

