% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-line-cutree.R
\name{geom_line_cutree}
\alias{geom_line_cutree}
\title{geom_line_cutree}
\usage{
geom_line_cutree(group, linetype = "dashed", offset = 0, ...)
}
\arguments{
\item{group}{output of cutree or number of subtree}

\item{linetype}{linetype}

\item{offset}{offset of the line}

\item{...}{additional parameters to set the line (e.g., color, size, etc.)}
}
\value{
line layer
}
\description{
geom_line_cutree
}
\examples{
d <- dist(USArrests)
hc <- hclust(d, "ave")
autoplot(hc) + geom_line_cutree(4)
}
