% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-rect-subtree.R
\name{geom_rect_subtree}
\alias{geom_rect_subtree}
\title{geom_rect_subtree}
\usage{
geom_rect_subtree(group = NULL, color = "red", ...)
}
\arguments{
\item{group}{output of cutree or number of subtree}

\item{color}{border color to highlight subtrees}

\item{...}{additional parameters pass to 'ggtree::hilight()'}
}
\value{
rect layer
}
\description{
geom_rect_subtree
}
\examples{
d <- dist(USArrests)
hc <- hclust(d, "ave")
autoplot(hc) + geom_rect_subtree(4)
}
\seealso{
\link[ggtree:geom-hilight]{geom_hilight};
}
