% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtreeDendro-package.R
\docType{package}
\name{ggtreeDendro-package}
\alias{ggtreeDendro}
\alias{ggtreeDendro-package}
\title{ggtreeDendro: Drawing 'dendrogram' using 'ggtree'}
\description{
Offers a set of 'autoplot' methods to visualize tree-like structures (e.g., hierarchical clustering and classification/regression trees) using 'ggtree'. You can adjust graphical parameters using grammar of graphic syntax and integrate external data to the tree.
}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Shuangbin Xu \email{xshuangbin@163.com} (\href{https://orcid.org/0000-0003-3513-5362}{ORCID}) [contributor]
  \item Chuanjie Zhang \email{959264868@qq.com} [contributor]
}

}
\keyword{internal}
