% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtreeDendro-package.R
\docType{package}
\name{ggtreeDendro}
\alias{ggtreeDendro}
\title{providing autoplot methods for many hierarchical clustering
results based on ggtree.}
\description{
This package implements a set of 'autoplot()' methods to display
tree structure. The output of it is a 'ggtree' object, which can be
annotated by adding layers using 'ggplot2' syntax. Users also can
integrate associated data to annotate the tree using 'ggtree' and
'ggtreeExtra' packages.
}
