% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna.R
\name{plot_wgcna}
\alias{plot_wgcna}
\title{plot_wgcna}
\usage{
plot_wgcna(x)
}
\arguments{
\item{x}{WGCNA::blockwiseModules() output}
}
\value{
ggtree object
}
\description{
plot wgcna result
}
\details{
Display WGCNA modules using ggtree. This function is similar to WGCNA::plotDendroAndColors()
}
\author{
Guangchuang Yu
}
