% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-color-subtree.R
\name{scale_color_subtree}
\alias{scale_color_subtree}
\alias{scale_colour_subtree}
\title{scale_color_subtree}
\usage{
scale_color_subtree(group)

scale_colour_subtree(group)
}
\arguments{
\item{group}{taxa group information}
}
\value{
updated tree view
}
\description{
scale tree color by subtree (e.g., output of cutree, kmeans, or other clustering algorithm)
}
\examples{
d <- dist(USArrests)
hc <- hclust(d, "ave")
autoplot(hc) + scale_color_subtree(3)
}
\author{
Guangchuang Yu
}
