% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{autoplot.linkage}
\alias{autoplot.hclust}
\alias{autoplot.dendrogram}
\alias{autoplot.agnes}
\alias{autoplot.diana}
\alias{autoplot.twins}
\title{Objects exported from other packages}
\usage{
\method{autoplot}{linkage}(object, layout = "dendrogram", ladderize = FALSE, ...)

\method{autoplot}{hclust}(object, layout = "dendrogram", ladderize = FALSE, ...)

\method{autoplot}{dendrogram}(object, layout = "dendrogram", ladderize = FALSE, ...)

\method{autoplot}{agnes}(object, layout = "dendrogram", ladderize = FALSE, ...)

\method{autoplot}{diana}(object, layout = "dendrogram", ladderize = FALSE, ...)

\method{autoplot}{twins}(object, layout = "dendrogram", ladderize = FALSE, ...)
}
\arguments{
\item{object}{input object}

\item{layout}{layout for plotting the tree}

\item{ladderize}{whether ladderize the tree (default FALSE)}

\item{...}{additional paramters that passed to ggtree}
}
\value{
ggtree object
}
\examples{
d <- dist(USArrests)
hc <- hclust(d, "ave")
autoplot(hc) + geom_tiplab()
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}
}}

